spring-cloud-ribbon/src/test/java/me/zhuyeqing/springclouddemo/HelloWorldTest.java
package me.zhuyeqing.springclouddemo;

import org.hamcrest.Matchers;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.HashMap;

import static org.junit.Assert.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringJUnit4ClassRunner.class)
@SpringBootTest(classes = RibbonApplication.class)
public class HelloWorldTest {
    private MockMvc mockMvc;
    @Before
    public void setUp(){
        mockMvc = MockMvcBuilders.standaloneSetup(new HelloWorld()).build();
    }

    @Test
    public void hello() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.get("/hello").param("name","World"))
                .andExpect(status().isOk()).andExpect(content().string(Matchers.equalTo("helloWorld")));

    }
    @Test
    public void testNullKey(){
        Object o = null;
        HashMap hs =new HashMap();
        hs.put(o,o);//when key is null, hashMap use 0 as key's hashcode
    }
}